#23 move Popup, Prompt and Welcome  to app folder
