+++ src/app/components/Popup/Popup.js
@@ -2,10 +2,10 @@ import React from "react";
 import { createHashHistory } from "history";
 import { HashRouter, Switch, Route } from "react-router-dom";
 
-import utils from "../lib/utils";
-import Home from "../app/pages/home";
-import Unlock from "../app/pages/unlock";
-import Loading from "../app/pages/loading";
+import utils from "../../../lib/utils";
+import Home from "../../pages/home";
+import Unlock from "../../pages/unlock";
+import Loading from "../../pages/loading";
 
 import "./styles.scss";
