+++ src/app/components/Popup/styles.scss
@@ -1,12 +1,13 @@
-@import "../styles/fonts";
-@import "../styles/reset";
-@import "../styles/variables";
+@import "../../styles/fonts";
+@import "../../styles/reset";
+@import "../../styles/variables";
+
+@import "~antd/dist/antd";
 
 body {
   color: $black;
   background-color: $greyWhite;
 }
-
 
 #popup {
   min-width: 350px;
